added documents, and seperated test (still to be moved into a unit test/integration framework)
