+++ src/stream/consumer.js
@@ -440,6 +440,8 @@ class Consumer extends EventEmitter {
    * 2. batch - event containing the batch of messages
    * 3. recursive - event to recursively call the recursive function - for internal use only!
    *
+   * @tutorial consumer
+   *
    * @fires Consumer#message
    * @fires Consumer#batch
    * @fires Consumer#recursive
@@ -589,150 +591,3 @@ class Consumer extends EventEmitter {
 exports.Consumer = Consumer
 // exports.CONSUMER_MODES = CONSUMER_MODES
 exports.ENUMS = ENUMS
-
-// exports.Stream = Consumer
-
-// TODO: TO BE MOVED INTO UNIT/INTEGRATION TESTS
-var testConsumer = async () => {
-  Logger.info('testConsumer::start')
-
-  var c = new Consumer(['test1'])
-
-  Logger.info('testConsumer::connect::start')
-  var connectionResult = await c.connect()
-  Logger.info('testConsumer::connect::end')
-  // c.connect().then(result => {
-  //   Logger.info(`Connected result=${result}`)
-  //
-  //   Logger.info('subscribing')
-  //   c.subscribe()
-  //
-  //   Logger.info('consuming')
-  //   c.consume()
-  // }).catch(error => Logger.error(error))
-
-  Logger.info(`Connected result=${connectionResult}`)
-
-  // Logger.info('testConsumer::subscribe::start')
-  // c.subscribe()
-  // Logger.info('testConsumer::subscribe::end')
-
-  Logger.info('testConsumer::consume::start1')
-
-  // c.consume((error, message) => {
-  //   if (error) {
-  //     Logger.info(`WTDSDSD!!! error ${error}`)
-  //   }
-  //   if (message) { // check if there is a valid message comming back
-  //     Logger.info(`Message Received by callback function - ${JSON.stringify(message)}`)
-  //
-  //     // lets check if we have received a batch of messages or single. This is dependant on the Consumer Mode
-  //     if (Array.isArray(message) && message.length != null && message.length > 0) {
-  //       message.forEach(msg => {
-  //         c.commitMessage(msg)
-  //       })
-  //     }
-  //   } else {
-  //     c.commitMessage(message)
-  //   }
-  //   setInterval(() => { Logger.info('WAITING 10000') }, 10000)
-  // })
-
-  c.consume((error, message) => {
-    return new Promise((resolve, reject) => {
-      if (error) {
-        Logger.info(`WTDSDSD!!! error ${error}`)
-        // resolve(false)
-        reject(error)
-      }
-      if (message) { // check if there is a valid message comming back
-        Logger.info(`Message Received by callback function - ${JSON.stringify(message)}`)
-        // lets check if we have received a batch of messages or single. This is dependant on the Consumer Mode
-        if (Array.isArray(message) && message.length != null && message.length > 0) {
-          message.forEach(msg => {
-            c.commitMessage(msg)
-          })
-        } else {
-          c.commitMessage(message)
-        }
-        resolve(true)
-      } else {
-        resolve(false)
-      }
-      // resolve(true)
-    })
-  })
-  Logger.info('testConsumer::consume::end1')
-
-  c.on('ready', arg => Logger.info(`onReady: ${JSON.stringify(arg)}`))
-  c.on('message', message => Logger.info(`onMessage: ${message.offset}, ${JSON.stringify(message.value)}`))
-  c.on('batch', message => Logger.info(`onBatch: ${JSON.stringify(message)}`))
-
-  Logger.info('testConsumer::end')
-}
-
-testConsumer()
-
-// //
-//
-// var testKafkaLib = () => {
-//   var consumer = new Kafka.KafkaConsumer({
-//     // 'debug': 'all',
-//     'metadata.broker.list': 'localhost:9092',
-//     'group.id': 'node-rdkafka-consumer-flow-example',
-//     'enable.auto.commit': false
-//   })
-//
-//   var topicName = 'test'
-//
-// // logging debug messages, if debug is enabled
-//   consumer.on('event.log', function (log) {
-//     console.log(log)
-//   })
-//
-// // logging all errors
-//   consumer.on('event.error', function (err) {
-//     console.error('Error from consumer')
-//     console.error(err)
-//   })
-//
-// // counter to commit offsets every numMessages are received
-//   var counter = 0
-//   var numMessages = 5
-//
-//   consumer.on('ready', function (arg) {
-//     console.log('consumer ready.' + JSON.stringify(arg))
-//
-//     consumer.subscribe([topicName])
-//     // start consuming messages
-//     consumer.consume()
-//   })
-//
-//   consumer.on('data', function (m) {
-//     counter++
-//
-//     // committing offsets every numMessages
-//     if (counter % numMessages === 0) {
-//       console.log('calling commit')
-//       consumer.commit(m)
-//     }
-//
-//     // Output the actual message contents
-//     console.log(JSON.stringify(m))
-//     console.log(m.value.toString())
-//   })
-//
-//   consumer.on('disconnected', function (arg) {
-//     console.log('consumer disconnected. ' + JSON.stringify(arg))
-//   })
-//
-// // starting the consumer
-//   consumer.connect()
-//
-// // stopping this example after 30s
-//   setTimeout(function () {
-//     consumer.disconnect()
-//   }, 30000)
-// }
-//
-// // testKafkaLib()
